wip: would've failed for larger datasets

+++ include/mstMerger.h
@@ -38,7 +38,7 @@ using FilterType = CQF<KeyObject>;
 typedef sdsl::bit_vector BitVector;
 typedef sdsl::rrr_vector<63> BitVectorRRR;
 typedef uint32_t colorIdType;
-typedef uint16_t weightType;
+typedef uint32_t weightType;
 
 struct Cost {
     uint64_t numSteps{0};
